Move metrics_collector to vppagent
